Rename Code Injection to Dependency Injection
