+++ .gitignore
@@ -27,3 +27,4 @@ node_modules/
 /playwright-report/
 /blob-report/
 /playwright/.cache/
+my_bank_app/
