+++ docs/docs/en/api/fastagency/api/dependency_injection/inject_params.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.api.code_injection.inject_params
+::: fastagency.api.dependency_injection.inject_params
