+++ tests/docs_src/test_import.py
@@ -25,7 +25,7 @@ MESOP_EXCLUDED_MODULES = {
     "docs_src.tutorials.giphy",
     "docs_src.tutorials.whatsapp",
     "docs_src.user_guide.runtimes.autogen.mesop",
-    "docs_src.user_guide.code_injection",
+    "docs_src.user_guide.dependency_injection",
 }
 
 # Mock Environment variables for Mesop Auth testing
